move ayon_api calls into lib
